% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_proforma.R
\name{create_proforma}
\alias{create_proforma}
\title{Create a ProForma strings from various identification tools}
\usage{
create_proforma(
  Sequence,
  Modifications,
  Tool,
  ConversionList = NULL,
  Scan = NULL,
  Protein = NULL
)
}
\arguments{
\item{Sequence}{(character) vector containing the "Sequence" column from MSPathFinder,
ProSight, or pTop. Required.}

\item{Modifications}{(character) vector containing the "Modifications" column
from MSPathFinder, the "PTMs" column from ProSight, or the "PTMs" column from
pTop. Must be the same length as Sequence. If there is no modification, use a blank string. Required.}

\item{Tool}{(character) indicates which tool was used. Acceptable inputs are "MSPathFinder".
"ProSight", or "pTop". Required.}

\item{ConversionList}{(list) used to convert ProSight modifications to
ProForma format. See example code and the pspecter modifications glossary with
read.csv(system.file("extdata", "Unimod_v20220602.csv", package = "pspecterlib")). Only
required for ProSight modifications.}

\item{Scan}{(numeric) vector containing the scan numbers. Must be the same length
as Sequence. Optional.}

\item{Protein}{(character) vector containing the protein names. Must be the same
length as Sequence. Optional.}
}
\value{
(data.frame) Returns a data.frame of scan numbers, proforma strings, and protein names.
Note that all unmodified sequences are included as well.
}
\description{
Derive proforma strings from outputs of MSPathFinder, ProSight, and
pTop. Note that TopPIC outputs ProForma strings so this step is
not required. For mzid files, use pull_modifications_from_mzid.
}
\examples{
\dontrun{

# MSPathFinder example with scan and protein
create_proforma(
   Sequence = c("GRGKTGGKARAKAKSRSSRAGLQFPVGRVHRL", "KKTRIIPRHLQLAIRNDEELNKLLGGVTIAY", "TEST"),
   Modifications = c("Methyl 8,Phospho 12,Phospho 22,DiMethyl 30", "TriMethyl 6,Phospho 15", ""),
   Tool = "MSPathFinder",
   Scan = c(3334, 3336, 3338),
   Protein = c("Protein33", "Protein45", "Protein47")
)

# ProSight example without scan and protein
create_proforma(
   Sequence = c("(49)M(37)SGRGKQG", "SG(67)RGKQGGKARAKAKSRSSRAG", "TEST"),
   Modifications = c("N-acetyl-L-methionine (49), O-phospho-L-serine (37)", "omega-N,omega-N'-dimethyl-L-arginine (67)", ""),
   Tool = "ProSight",
   ConversionList = list("N-acetyl-L-methionine" = "Acetyl", "O-phospho-L-serine" = "Phospho",
                         "omega-N,omega-N'-dimethyl-L-arginine" = "DiMethyl")
)

# pTop Example
create_proforma(
   Sequence = c("SGRGKGGKGLGKGGAKRHRKVLRDNIQGITKPAIRRL", "TEST", "PEPSRSTPAPKKGSKKAITKAQKKDGKKRKRGRKESYSIYV"),
   Modifications = c("(20)Dimethyl[K];(16)Acetyl[K];(0)Acetyl[AnyN-term];", "", "(20)Dimethyl[K];"),
   Tool = "pTop"
)

}
}
