% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks.R
\name{filter_peaks}
\alias{filter_peaks}
\title{Filter a pspecterlib peak_data object}
\usage{
filter_peaks(PeakData, MZRange, NoiseFilter = 1)
}
\arguments{
\item{PeakData}{A peak_data object generated by get_peak_data (from mzML or raw),
or make_peak_data in pspecterlib. Required.}

\item{MZRange}{A range of MZ values to filter the data by.
It is highly recommended that users visualize the spectra first to determine
a reasonable cutoff range. Required.}

\item{NoiseFilter}{An abundance (every peak is scaled to the largest peak) cutoff
for peaks. A reasonable value should be in the 0.01 - 5.0\% range. Default is 1\%.}
}
\value{
A peak_data object that has been filtered.
}
\description{
For best results in the ProteoMatch algorithm, it's best to filter
by a specific M/Z range and a noise threshold.
}
\examples{
\dontrun{

# Download an example file
tmpdir <- tempdir()
TopDownFile <- "https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/TopDown/TopDown.mzML"
download.file(TopDownFile, file.path(tmpdir, tail(unlist(strsplit(TopDownFile, "/")), 1)))

# Read with pspecterlib (mzR on the backend)
TD_ScanMetadata <- pspecterlib::get_scan_metadata(MSPath = file.path(tmpdir, "TopDown.mzML"))

# Extract MS1 peak data
PeakData <- pspecterlib::get_peak_data(TD_ScanMetadata, 5684)

# Plot peak data
pspecterlib::annotated_spectrum_plot(PeakData = PeakData, Interactive = TRUE)

# Cutoff at 2.5\% and 97.5\% percentile ranges
quantile(PeakData$`M/Z`, c(0.025, 0.975))

# Set a cutoff of MZ range
filter_peaks(
   PeakData = PeakData,
   MZRange = c(640, 1550),
   NoiseFilter = 2.5
)

}

}
