% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomatchms_trelliscope.R
\name{isomatchms_trelliscope}
\alias{isomatchms_trelliscope}
\title{Generate a trelliscope display for all Biomolecule to MS1 Matches}
\usage{
isomatchms_trelliscope(
  PeakData,
  Ms1Match,
  Path = file.path(.getDownloadsFolder(), "Ms1Match", "Trelliscope"),
  MinCorrelationScore = 0.7,
  Window = 2
)
}
\arguments{
\item{PeakData}{(peak_data object) A pspecterlib or data.table with "M/Z" and "Intensity". Required.}

\item{Ms1Match}{(IsoMatchMS_MatchedPeaks object) object from match_full_seq_ms1. Required.}

\item{Path}{(character) The base directory of the trelliscope application. Default is Downloads/Ms1Match.}

\item{MinCorrelationScore}{(numeric) The minimum correlation score to plot. Default is 0.7.}

\item{Window}{(numeric) The -/+ m/z value on either side of the matched spectra plot. Default is 2 m/z.}
}
\value{
An html trelliscope display
}
\description{
\href{https://github.com/hafen/trelliscopejs}{Trelliscope} allows
users to filter and sort plots based on cognostic variables. For this dataset,
the plots are generated with the plot_Ms1Match function, and the cognostics
are: Identifiers, Absolute Relative Error, Correlation, Charge, Biomolecules, and ID.
}
\examples{
\dontrun{

# Run two examples with two charge states
MolForms_Test <- calculate_molform(
  Biomolecule = c("M.SS[Methyl]S.V", "M.SS[6]S[7].V"),
  BioType = "ProForma",
  Identifiers = c("Test1", "Test2"),
  Charge = 1:2
)

# Generate some experimental peak data to match
PeakData <- pspecterlib::make_peak_data(
  MZ = c(147.5684, 148.0699, 148.5708, 149.0721, 149.5731,
         294.1296, 295.1325, 296.1343, 297.1369, 298.1390),
  Intensity = c(868.3680036, 110.9431876, 18.7179196, 1.7871629, 0.1701294,
                868.3680036, 110.9431876, 18.7179196, 1.7871629, 0.1701294)
)

# Run algorithm
IsoMatch <- match_biomolecule_to_ms1(
  PeakData = PeakData,
  MatchingAlgorithm = "closest peak",
  MolecularFormula = MolForms_Test,
  IsotopeRange = c(3, 20)
)

# Make the trelliscope display
isomatchms_trelliscope(PeakData = PeakData, Ms1Match = IsoMatch)

}


}
