% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_biomolecule_to_ms1.R
\name{match_biomolecule_to_ms1}
\alias{match_biomolecule_to_ms1}
\title{Matches experimental peaks to calculated isotope profile}
\usage{
match_biomolecule_to_ms1(
  PeakData,
  MolecularFormulas,
  AbundanceThreshold = 50,
  PPMThreshold = 10,
  IsotopeMinimum = 3,
  IsotopeAlgorithm = "Rdisop",
  IncludeAll = FALSE
)
}
\arguments{
\item{PeakData}{(peak_data object) from pspecterlib, made with "make_peak_data" or
extracted from a raw or mzML file with "get_peak_data." Use of centroided data
is recommended, but not required.}

\item{MolecularFormulas}{(IsoMatchMS_MolForm object) object with Molecular Formulas,
Mass Shifts, and Charges.}

\item{AbundanceThreshold}{(numeric) The +/- percent abundance an isotope peak
can vary and still be considered a match. For example, if the true isotope peak
is 3 but the measured value is 4, an AbundanceThreshold of 50\% will consider
that peak a true match if it ranges between 1.5 - 4.5. Default is 50.}

\item{PPMThreshold}{(numeric) The window in PPM to search for peaks around the true peak. Required. Default is 10 ppm.}

\item{IsotopeMinimum}{(numeric) The minimum number of isotopes to consider. Default is 3.}

\item{IsotopeAlgorithm}{(character) "isopat" uses the isopat package to calculate isotopes,
while "Rdisop" uses the Rdisop package. Though more accurate, Rdisop has been known
to crash on Windows computers when called iteratively more than 1000 times.
Default is Rdisop, though isopat is an alternative.}

\item{IncludeAll}{(logical) Return non-matches along with matches. Default is FALSE.}
}
\value{
(IsoMatchMS_MatchedPeaks object) A IsoMatchMS_MatchedPeaks object, which is a data.table containing the
Identifier, Adduct, M/Z, Intensity, Isotope, M/Z Search Window, Experimental M/Z,
Experimental Intensity, PPM Error, Absolute Relative Error, Correlation,
Charge, Biomolecule, and an ID.
}
\description{
Returns the "IsoMatchMS_MatchedPeaks" object with peaks that
have matched to the isotope profile. This object can be passed to downstream
plotting functions.
}
\details{
The data.table outputted by this function contains 12 columns
\tabular{ll}{
Identifiers \tab The provided biomolecule identifier \cr
\tab \cr
Adduct Mass \tab The provided adduct masses \cr
\tab \cr
Adduct Names \tab The provided adduct names \cr
\tab \cr
M/Z \tab The calculated M/Z of a particular isotope \cr
\tab \cr
Mass Shift \tab The mass shifts summed in the ProForma string \cr
\tab \cr
Monoisotopic Mass \tab The monoisotopic mass for the isotope profile \cr
\tab \car
Abundance \tab The measured abundance value \cr
\tab \cr
Isotope \tab The isotope number "n" in M+n \cr
\tab \cr
M/Z Search Window \tab Given the ppm, the M/Z Search Window is the window in which peaks are searched and matched \cr
\tab \cr
M/Z Experimental \tab The M/Z of the experimental peaks \cr
\tab \cr
Intensity Experimental \tab The intensity of the experimental peak \cr
\tab \cr
Abundance Experimental \tab The abundance of the experimental peak
\tab \cr
PPM Error \tab How far off the calculated peak is from the experimental peak. Only peaks within the ppm threshold are kept \cr
\tab \cr
Absolute Relative Error \tab The sum of the absolute relative difference of calculated and experimental peak intensities. \cr
\tab \cr
Pearson Correlation \tab The pearson correlation of calculated and experimental peak intensities \cr
\tab \cr
Charge \tab The provided charges \cr
\tab \cr
Biomolecules \tab The provided biomolecule \cr
\tab \cr
Molecular Formula \tab The calculated molecular formula \cr
\tab \cr
Most Abundant Isotope \tab The most abundant isotope, if calculated in calculate_molform \cr
\tab \cr
ID \tab A unique ID for each Proteoform, Protein, and Charge combination used in plotting functions \cr
\tab \cr
}
}
\examples{
\dontrun{

# Run two examples with two charge states
MolForms_Test <- calculate_molform(
   Biomolecules = c("M.SS[Methyl]S.V", "M.SS[6]S[7].V"),
   BioType = "ProForma",
   Identifiers = c("Test1", "Test2"),
   Charge = 1:2
)

# Generate some experimental peak data to match
PeakData <- pspecterlib::make_peak_data(
   MZ = c(294.1296, 295.1325, 296.1343, 297.1369, 298.1390),
   Intensity = c(868.3680036, 110.9431876, 18.7179196, 1.7871629, 0.1701294)
)

# Run algorithm
match_biomolecule_to_ms1(
    PeakData = PeakData,
    MolecularFormula = MolForms_Test,
    IsotopeMinimum = 2
)

}

}
