% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomatch_trelliscope.R
\name{proteomatch_trelliscope}
\alias{proteomatch_trelliscope}
\title{Generate a trelliscope display for all Proteoform to MS1 Matches}
\usage{
proteomatch_trelliscope(
  PeakData,
  Ms1Match,
  Path = file.path(.getDownloadsFolder(), "Ms1Match", "Trelliscope"),
  MinCorrelationScore = 0.7,
  Window = 2
)
}
\arguments{
\item{PeakData}{A pspecterlib peak_data object or data.table with "M/Z" and "Intensity". Required.}

\item{Ms1Match}{A ProteoMatch_MatchedPeaks class object from match_full_seq_ms1. Required.}

\item{Path}{The base directory of the trelliscope application. Default is Downloads/Ms1Match.}

\item{MinCorrelationScore}{The minimum correlation score to plot. Default is 0.7.}

\item{Window}{The -/+ m/z value on either side of the matched spectra plot. Default is 2 m/z.}
}
\value{
An html trelliscope display
}
\description{
\href{https://github.com/hafen/trelliscopejs}{Trelliscope} allows
users to filter and sort plots based on cognostic variables. For this dataset,
the plots are generated with the plot_Ms1Match function, and the cognostics
are: Protein, Absolute Relative Error, Correlation, Charge, Proteoform, and ID.
}
\examples{
\dontrun{

# Run two examples with two charge states
MolForms_Test <- calculate_molform(
  Proteoform = c("M.SS[Methyl]S.V", "M.SS[6]S[7].V"),
  Protein = c("Test1", "Test2"),
  Charge = 1:2
)

# Generate some experimental peak data to match
PeakData <- pspecterlib::make_peak_data(
  MZ = c(147.5684, 148.0699, 148.5708, 149.0721, 149.5731,
         294.1296, 295.1325, 296.1343, 297.1369, 298.1390),
  Intensity = c(868.3680036, 110.9431876, 18.7179196, 1.7871629, 0.1701294,
                868.3680036, 110.9431876, 18.7179196, 1.7871629, 0.1701294)
)

# Run algorithm
ProteoMatch <- match_proteoform_to_ms1(
  PeakData = PeakData,
  MolecularFormula = MolForms_Test,
  IsotopeRange = c(3, 20)
)

# Make the trelliscope display
proteomatch_trelliscope(PeakData = PeakData, Ms1Match = ProteoMatch)

}


}
