% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_modifications_from_mzid.R
\name{pull_modifications_from_mzid}
\alias{pull_modifications_from_mzid}
\title{Pull proforma strings}
\usage{
pull_modifications_from_mzid(IDPath)
}
\arguments{
\item{IDPath}{The path to the mzid file. Required.}
}
\value{
Returns a data.frame of scan numbers, proforma strings, and protein names.
Note that all unmodified sequences are included as well.
}
\description{
Pull modifications from MS-GF+ and MSPathFinder files (mzid).
}
\examples{
\dontrun{

# Create a temporary directory and copy example data there
tmpdir <- tempdir()

files <- c(
"https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/BottomUp/BottomUp.mzid",
"https://raw.githubusercontent.com/EMSL-Computing/PSpecteR/master/pspecter_container/TestFiles/TopDown/TopDown.mzid"
)

for (file in files) {
 download.file(file, file.path(tmpdir, tail(unlist(strsplit(file, "/")), 1)))
}

# Download a MS-GF+ example
pull_modifications_from_mzid(file.path(tmpdir, "BottomUp.mzid"))

# Download an MSPathFinder example
pull_modifications_from_mzid(file.path(tmpdir, "TopDown.mzid"))

}
}
