% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isomatchms.R
\name{run_isomatchms}
\alias{run_isomatchms}
\title{Run the IsoMatchMS pipeline with a list of PTMs, summed peak data, and algorithm settings}
\usage{
run_isomatchms(
  Biomolecules,
  BioType,
  SummedSpectra,
  SettingsFile,
  Path = file.path(.getDownloadsFolder(), "Ms1Match"),
  Identifiers = NULL,
  Messages = TRUE
)
}
\arguments{
\item{Biomolecules}{A vector of Molecular Formulas or ProForma strings (i.e. "M.AA\link{Acetyl⁠}⁠AA\link{3.2⁠}⁠.V").
ProForma strings can be pulled from mzid files (MS-GF+, MSPathFinder) with pull_modifications_from_mzid,
or made with create_proforma for MSPathFinder, ProSight, and pTop modifications. TopPIC proteoforms
are provided as ProForma strings. Required.}

\item{BioType}{A string indicating whether the Biomolecules are "ProForma" strings or "Molecular Formula". Required.}

\item{SummedSpectra}{A pspecterlib peak_data object of all summed MS1 spectra. This
can be generated with an mzML file and the sum_ms1_spectra function, pulled from
a summed mzML with pspecterlib::get_peak_data, or simply created with
pspecterlib::make_peak_data. Required.}

\item{SettingsFile}{Path to a xlsx file with all user-set parameters. For examples,
try xlsx::read.xlsx(system.file("extdata", "Intact_Protein_Defaults.xlsx", package = "IsoMatchMS"), 1).
Required.}

\item{Path}{The base directory of the output trelliscope application. Default is Downloads/Ms1Match.}

\item{Identifiers}{A vector of identifiers for each biomolecule (i.e. protein, glycan, etc.). Optional.}

\item{Messages}{A TRUE/FALSE indicating whether messages should be printed.}
}
\value{
\tabular{ll}{
MolecularFormulas \tab A CSV with all calculated molecular formulas from the ProForma strings. \cr
\tab \cr
FilteredPeaks \tab A CSV with the final spectrum used after filtering \cr
\tab \cr
Matched_Isotope_Distributions \tab A CSV with all matched isotope distributions, regardless of quality \cr
\tab \cr
Trelliscope_Display \tab Visualization of all matched isotope distributions above the quality threshold \cr
\tab \cr
}
}
\description{
A wrapper for quickly running all the steps of the IsoMatchMS pipeline
}
\details{
Before running the main algorithm, you will need Molecular Formulas (or ProForma strings)
and a summed spectra. There are several ways to get ProForma strings, including
pull_modifications_from_mzid and create_proforma. Spectra may be summed with sum_ms1_spectra,
or uploaded from an mzML (see pspecterlib::get_scan_metadata and pspecterlib::get_peak_data) or
created from a csv (see pspecterlib::make_peak_data).

The algorithm is run in 4 main steps:
\enumerate{
\item Molecular Formulas and masses are calculated with calculate_molform
\item Input summed spectra are filtered with filter_peaks
\item Isotopic distributions are calculated with the molecular formulas, and matched
to the filtered spectra
\item All high scoring matches are visualized with a trelliscope display
}

Documentation for example intact proteomic data can be found \href{https://www.sciencedirect.com/science/article/pii/S1535947622002997}{here}

Documentation for example digested proteomic data can be found \href{https://google.com}{here}

Documentation for example glycan data can be found \href{https://google.com}{here}
}
\examples{
\dontrun{

####################
## SHORT EXAMPLES ##
####################

ToDo

###################
## FULL EXAMPLES ##
###################

## INTACT PROTEOMICS EXAMPLE ##

# Extract the summed peak data with any tool (here, pspecterlib, which uses mzR)
MSPath <- system.file("extdata", "Intact_Protein_Summed_MS1.mzML", package = "IsoMatchMS")
ScanMetadata <- pspecterlib::get_scan_metadata(MSPath)
PeakData <- pspecterlib::get_peak_data(ScanMetadata, ScanNumber = 1, MinAbundance = 0.1)

# Now, get the molecular formulas (required) and protein names (optional)
ProteinData <- read.csv(system.file("extdata", "Intact_Proteins_List.csv", package = "IsoMatchMS"), header = T)
Proteoform <- ProteinData$Proteoform
ProteinNames <- ProteinData$Protein.accession

# Get the settings file path
Intact_Settings_Path <- system.file("extdata", "Intact_Protein_Defaults.xlsx", package = "IsoMatchMS")

# Now, run the main pipeline 
run_isomatchms(
    Biomolecule = Proteoform,
    BioType = "ProForma",
    SummedSpectra = PeakData, 
    SettingsFile = Intact_Settings_Path, 
    Path = file.path(.getDownloadsFolder(), "Intact_Example"),
    Identifiers = ProteinNames,
    Messages = TRUE
)

## DIGESTED PROTEOMICS EXAMPLE ##

# Pull summed peak data from CSV
Peaks <- read.csv(system.file("extdata", "Peptides_PeakData.csv", package = "IsoMatchMS"))
PeakData <- pspecterlib::make_peak_data(MZ = Peaks$M.Z, Intensity = Peaks$Intensity)

# Now, get the molecular formulas (required) and protein names (optional)
ProteinData <- read.csv(system.file("extdata", "Peptides_List.csv", package = "IsoMatchMS"))
ProForma <- ProteinData$Proteoform
ProteinNames <- ProteinData$Protein

# Get the settings file path
Digested_Settings_Path <- system.file("extdata", "Peptides_Defaults.xlsx", package = "IsoMatchMS")

# Now, run the main pipeline 
run_isomatchms(
    Biomolecule = ProForma,
    BioType = "ProForma",
    SummedSpectra = PeakData, 
    SettingsFile = Digested_Settings_Path, 
    Path = file.path(.getDownloadsFolder(), "Digested_Example"),
    Identifiers = ProteinNames,
    Messages = TRUE
)

}

}
