% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_ms1_spectra.R
\name{sum_ms1_spectra}
\alias{sum_ms1_spectra}
\title{Sum MS1 Spectra}
\usage{
sum_ms1_spectra(
  mzMLPath,
  PPM_Round = 5,
  MinimumAbundance = 0.01,
  Messages = FALSE
)
}
\arguments{
\item{mzMLPath}{The path to the mzML file. Required.}

\item{PPM_Round}{Number to round to in PPM. Default is 5.}

\item{MinimumAbundance}{The minimum abundance value to keep. Acceptable values
range from 0 to 100. Default is 0.01.}

\item{Messages}{A TRUE/FALSE to indicate whether help messages should be printed or not.
Default is TRUE.}
}
\value{
A peak_data object with summed MS1 spectra. This object can be directly
inputted into run_proteomatch.
}
\description{
Reads an mzML file, extracts all MS1 spectra, and sums the abundances
together at a PPM threshold
}
\details{
Allows users to easily sum spectra for the main ProteoMatch pipeline function
run_proteomatch. Values are rounded to the nearest 5 ppm. After summing, abundances
less than 0.01 are tossed.
}
\examples{
\dontrun{


}
}
