% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_biomolecule_to_ms1.R
\name{match_biomolecule_to_ms1}
\alias{match_biomolecule_to_ms1}
\title{Matches experimental peaks to calculated isotope profile}
\usage{
match_biomolecule_to_ms1(
  PeakData,
  MolecularFormulas,
  MatchingAlgorithm,
  AbundanceThreshold = 0.5,
  PPMThreshold = 10,
  IsotopeRange = c(5, 20)
)
}
\arguments{
\item{PeakData}{(peak_data object) from pspecterlib, made with "make_peak_data" or
extracted from a raw or mzML file with "get_peak_data." Use of centroided data
is recommended, but not required.}

\item{MolecularFormulas}{(IsoMatchMS_MolForm object) object with Molecular Formulas,
Mass Shifts, and Charges.}

\item{MatchingAlgorithm}{(character) Either "closest peak" or "highest abundance" where the "closest
peak" implementation chooses the peak closest to the true M/Z value within the PPM window
and "highest abundance" chooses the highest intensity peak within the PPM window. "closest peak"
is recommended for peaks that have been peak picked with an external tool,
and "highest abundance" is recommended for noisy datasets or those with many peaks.}

\item{PPMThreshold}{(numeric) The window in PPM to search for peaks around the true peak. Required. Default is 10 ppm.}

\item{IsotopeRange}{(numeric) The minimum and maximum number of isotopes to consider. Default is c(3,20).}

\item{MinAbundance}{(numeric) The minimum abundance (calculated intensity) permitted
to be matched. Default is 0.1, which is 0.1\%.}
}
\value{
(IsoMatchMS_MatchedPeaks object) A IsoMatchMS_MatchedPeaks object, which is a data.table containing the
Identifier, Adduct, M/Z, Intensity, Isotope, M/Z Search Window, Experimental M/Z,
Experimental Intensity, PPM Error, Absolute Relative Error, Correlation,
Charge, Biomolecule, and an ID.
}
\description{
Returns the "IsoMatchMS_MatchedPeaks" object with peaks that
have matched to the isotope profile. This object can be passed to downstream
plotting functions.
}
\details{
The data.table outputted by this function contains 12 columns
\tabular{ll}{
Identifier \tab The provided biomolecule identifier \cr
\tab \cr
Adduct \tab The provided adduct \cr
\tab \cr
M/Z \tab The calculated M/Z of a particular isotope \cr
\tab \cr
Intensity \tab The calculated relative intensity of a particular isotope \cr
\tab \cr
Isotope \tab The isotope number "n" in M+n \cr
\tab \cr
M/Z Search Window \tab Given the ppm, the M/Z Search Window is the window in which peaks are searched and matched \cr
\tab \cr
M/Z Experimental \tab The M/Z of the matched from the experimental peaks \cr
\tab \cr
Intensity Experimental \tab The intensity of the matched peak from the experimental peaks \cr
\tab \cr
PPM Error \tab How far off the calculated peak is from the experimental peak. Only peaks within the ppm threshold are kept \cr
\tab \cr
Absolute Relative Error \tab The sum of the absolute relative difference of calculated and experimental peak intensities. \cr
\tab \cr
Correlation \tab The cosine correlation of calculated and experimental peak intensities \cr
\tab \cr
Charge \tab The provided charges \cr
\tab \cr
Biomolecule \tab The provided biomolecule \cr
\tab \cr
ID \tab A unique ID for each Proteoform, Protein, and Charge combination used in plotting functions \cr
\tab \cr
}
}
\examples{
\dontrun{

# Run two examples with two charge states
MolForms_Test <- calculate_molform(
   Biomolecules = c("M.SS[Methyl]S.V", "M.SS[6]S[7].V"),
   BioType = "ProForma",
   Identifiers = c("Test1", "Test2"),
   Charge = 1:2
)

# Generate some experimental peak data to match
PeakData <- pspecterlib::make_peak_data(
   MZ = c(294.1296, 295.1325, 296.1343, 297.1369, 298.1390),
   Intensity = c(868.3680036, 110.9431876, 18.7179196, 1.7871629, 0.1701294)
)

# Run algorithm
match_biomolecule_to_ms1(
   PeakData = PeakData,
   MolecularFormula = MolForms_Test,
   MatchingAlgorithm = "closest peak",
   IsotopeRange = c(3,20)
)

}

}
